Move stub code to own file.
